requirements.txt
1.Cowsay
2.Inflect
3.PyFiglet
4.tabulate